% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODO.R
\name{buildTODO}
\alias{buildTODO}
\title{Build TODO list from package directory}
\usage{
buildTODO()
}
\arguments{
\item{taxaList}{A vector of species names to be validated}

\item{dictionary}{A vector of valid species names}

\item{synTable}{A dataframe with current names paired with previous synonyms}
}
\value{
a dataframe of mis-matched names and their suggested taxonomy.
}
\description{
Addin scans all .R and .Rmd files in the poject directory for
  'TODO' comments TODO items are grouped by file name and ordered by line
  number, then writen to a seperate file: TODO.txt This file is overwritten
  each time the addin is executed.
}

